Add jsui system, add flexbox layout module
jsui has a basic loadModule interface that expects an object with a 'dom_handlers' or 'class_handlers' object. This object contains key=>value pairs of HTML tag or class names pointing to a function to be called when that tag or class is created.

A flexbox module, 'Boxr', has been provided that implements flexbox functionality via the 'box', 'vbox', and 'hbox' custom HTML tags. These tags can contain 'size' and 'grip' attributes. The 'size' attribute designates the flexGrow attribute and the 'grip' attribute designates the type of gripper (left,right,up,down) to use for dynamically resizing the flexbox element.