Separate modules into files, test code for Tree module
Modules now exist as `lib/module_....js` -- also included is a compile.sh script that simply concats `jsui.js` and any passed modules into a jsui.js file in the root directory. Usage would be similar to `sh compile.sh boxr tabs tree`, resulting in a jsui.js file containing those.

Test code is in place for a Tree module, but it is not implemented proper as of yet.

Styling has been somewhat changed on boxr and tabs, however a unified styling has not been reached properly as of yet.