Add Grid module, appending to DOM now triggers class/tag handlers
The Grid module basically creates a fixed-cell size table based on the
grid element's `rows`, `cols`, `cellwidth`, and `cellheight` attributes.

It emits a `cell-click` Event with `cell`, `x`, and `y` properties when
a cell is clicked.

Class and tag handlers will now be properly triggered when the
appropriate class or tag is added to the DOM. All of the added element's
children will also be processed.

At the moment, we are using a ktkInit attribute to denote if the element
has been processed.